Fix small MC sampling bugs that led to NaNs in rare cases when the PDF being sampled was zero in some buckets (and thus the CDF values at the edges of each bucket were the same).
In particular, it was happening that sometimes a sample value of u=0 would be passed into the sampling routines when the CDF array had values like {0, 0, 0, .1, .2, ... }.  The old code would pick the first bucket, from [0,0], and then compute a NaN value for the du offset (0/(0-0) == NaN).  The new code instead picks the bucket [0,.1] instead and all the math works out as desired.