Made the parser handle relative filenames better.
Filenames are resolved relative to the directory containing the current parser input file.

To do this, I made ParamSet instances store the directory that relative filenames should be resolved against. Then I added the ResolveFilename method which does the resolution (with correct handling of absolute filenames); and a FindOneFilename method which looks up a string and calls ResolveFilename on it. I also add a FindFilename method to the TextureParams class, as a convenience.

I then modified the classes (some of the texture types, some of the lights) which grab filenames from a ParamSet to use the FindOneFilename/FindFilename function instead of FindString.

Limitations:
- I don't think this will work for included files yet.
- The change won't work on Windows (it uses libgen.h and it looks for '/' at the start of a filename to tell whether it's absolute or not).