Includes are now handled by the parser, not the lexer.
Previously the lexer was handling includes so that they were transparent to the parser. This was causing a problem with the handling of relative filenames in any statement immediately preceding an include. The parser has to advance into the included file to find the end of the statement, so by the time it initialises the paramset for the statement the current_file has changed.

The fix is to move the include handling out of the lexer and into the parser. Now an include statement is parsed like any other statement and it's semantic action calls include_push. This means that an Include will terminate the preceding statement before we open the included file.

The downside is that certain things which worked before won't anymore. You won't be able to Include a statement fragment, so something like this

	Shape "trianglemesh" "point P" [ Include(points.txt) ]

won't work anymore. I don't know how much of a problem that will be.