Remove #error directive for g++ 4.8.2
The version number is not enough to know that std::nth_element is buggy.
GCC 4.8.2 is the compiler for both Ubuntu 14.04 and CentOS 7, and both
distributions have patched libstdc++ to fix GCC Bug #58800.

At this point, the #error directive is more likely to cause problems
than prevent problems.

Tested with killeroo-gold.pbrt on Ubuntu 14.04 and CentOS 7.
This reverts commit a7bf7eaea319f7927ceb478cd3df95545e5e1d14.