Fix bug in GridAccelerator for rays with zero direction components.
The grid accelerator DDA was carefully written to allow for zero values of
ray direction components, taking advantage of the fact that for x>0,
x/0. == Infinity, and Infinity behaves in reasonable expected ways.

However... IEEE float also has a negative zero value. If we happened to get
a ray with a negative zero component, things would go bad, since we'd
decide that the next t value along that dimension was the one we should
take.

Reported by Mike Phillips.